Update build_sam_hq.py to also run on only cpu (#444)
Currently we get "RuntimeError: Attempting to deserialize object on a CUDA device but torch.cuda.is_available() is False. If you are running on a CPU-only machine, please use torch.load with map_location=torch.device('cpu') to map your storages to the CPU." error while running on cpu device.